Fix bug w/ lane index > 32-bit
Also move SIMD lane count function into opcode.h

+++ src/type-checker.h
@@ -87,7 +87,7 @@ class TypeChecker {
   Result OnSelect();
   Result OnSetGlobal(Type);
   Result OnSetLocal(Type);
-  Result OnSimdLaneOp(Opcode, int);
+  Result OnSimdLaneOp(Opcode, uint64_t);
   Result OnSimdShuffleOp(Opcode, v128);
   Result OnStore(Opcode);
   Result OnTeeLocal(Type);
